+++ src/controllers/SettingsController.php
@@ -17,7 +17,9 @@ use nystudio107\seomatic\helpers\Schema as SchemaHelper;
 use nystudio107\seomatic\helpers\ArrayHelper;
 use nystudio107\seomatic\helpers\DynamicMeta as DynamicMetaHelper;
 use nystudio107\seomatic\helpers\ImageTransform as ImageTransformHelper;
+use nystudio107\seomatic\helpers\PluginTemplate;
 use nystudio107\seomatic\models\MetaBundle;
+use nystudio107\seomatic\models\MetaScript;
 use nystudio107\seomatic\models\MetaScriptContainer;
 use nystudio107\seomatic\services\FrontendTemplates;
 use nystudio107\seomatic\services\MetaBundles;
@@ -220,7 +222,7 @@ class SettingsController extends Controller
      * @throws NotFoundHttpException
      * @throws \yii\web\ForbiddenHttpException
      */
-    public function actionGlobal(string $subSection = 'general', string $siteHandle = null, $loadFromSiteHandle = null): Response
+    public function actionGlobal(string $subSection = 'general', string $siteHandle = null, $loadFromSiteHandle = null, $editedMetaBundle = null): Response
     {
         $variables = [];
         $siteId = $this->getSiteIdFromHandle($siteHandle);
@@ -275,6 +277,9 @@ class SettingsController extends Controller
         $siteIdToLoad = $loadFromSiteHandle === null ? (int)$variables['currentSiteId'] : $loadFromSiteId;
         // Load the metabundle
         $metaBundle = Seomatic::$plugin->metaBundles->getGlobalMetaBundle($siteIdToLoad);
+        if ($editedMetaBundle) {
+            $metaBundle = $editedMetaBundle;
+        }
         Seomatic::$previewingMetaContainers = false;
         if ($metaBundle !== null) {
             $variables['metaGlobalVars'] = clone $metaBundle->metaGlobalVars;
@@ -321,7 +326,7 @@ class SettingsController extends Controller
      * @throws \yii\web\BadRequestHttpException
      * @throws \craft\errors\MissingComponentException
      */
-    public function actionSaveGlobal(): Response
+    public function actionSaveGlobal()
     {
         $this->requirePostRequest();
         $request = Craft::$app->getRequest();
@@ -335,6 +340,7 @@ class SettingsController extends Controller
         // Set the element type in the template
         $elementName = '';
 
+        $hasErrors = false;
         // The site settings for the appropriate meta bundle
         Seomatic::$previewingMetaContainers = true;
         $metaBundle = Seomatic::$plugin->metaBundles->getGlobalMetaBundle($siteId);
@@ -353,14 +359,33 @@ class SettingsController extends Controller
             $robotsContainer = $templateContainers[FrontendTemplates::ROBOTS_TXT_HANDLE];
             if ($robotsContainer !== null && \is_array($robotsTemplate)) {
                 $robotsContainer->setAttributes($robotsTemplate);
+                if (!$robotsContainer->validate()) {
+                    $hasErrors = true;
+                }
             }
             $humansContainer = $templateContainers[FrontendTemplates::HUMANS_TXT_HANDLE];
             if ($humansContainer !== null && \is_array($humansTemplate)) {
                 $humansContainer->setAttributes($humansTemplate);
+                if (!$humansContainer->validate()) {
+                    $hasErrors = true;
+                }
             }
             $adsContainer = $templateContainers[FrontendTemplates::ADS_TXT_HANDLE];
             if ($adsContainer !== null && \is_array($adsTemplate)) {
                 $adsContainer->setAttributes($adsTemplate);
+                if (!$adsContainer->validate()) {
+                    $hasErrors = true;
+                }
+            }
+
+            if ($hasErrors) {
+                Craft::$app->getSession()->setError(Craft::t('app', "Couldn't save settings due to a Twig error."));
+                // Send the redirect back to the template
+                Craft::$app->getUrlManager()->setRouteParams([
+                    'editedMetaBundle' => $metaBundle,
+                ]);
+
+                return null;
             }
 
             Seomatic::$plugin->metaBundles->syncBundleWithConfig($metaBundle, true);
@@ -821,7 +846,7 @@ class SettingsController extends Controller
      * @throws NotFoundHttpException
      * @throws \yii\web\ForbiddenHttpException
      */
-    public function actionTracking(string $subSection = 'gtag', string $siteHandle = null, $loadFromSiteHandle = null): Response
+    public function actionTracking(string $subSection = 'gtag', string $siteHandle = null, $loadFromSiteHandle = null, $editedMetaBundle = null): Response
     {
         $variables = [];
         // Get the site to edit
@@ -839,6 +864,9 @@ class SettingsController extends Controller
         $siteIdToLoad = $loadFromSiteHandle === null ? (int)$variables['currentSiteId'] : $loadFromSiteId;
         // Load the metabundle
         $metaBundle = Seomatic::$plugin->metaBundles->getGlobalMetaBundle($siteIdToLoad);
+        if ($editedMetaBundle) {
+            $metaBundle = $editedMetaBundle;
+        }
         Seomatic::$previewingMetaContainers = false;
         if ($metaBundle !== null) {
             $variables['scripts'] = Seomatic::$plugin->metaBundles->getContainerDataFromBundle(
@@ -894,7 +922,7 @@ class SettingsController extends Controller
      * @throws \yii\web\BadRequestHttpException
      * @throws \craft\errors\MissingComponentException
      */
-    public function actionSaveTracking(): Response
+    public function actionSaveTracking()
     {
         $this->requirePostRequest();
         $request = Craft::$app->getRequest();
@@ -905,12 +933,14 @@ class SettingsController extends Controller
         Seomatic::$previewingMetaContainers = true;
         $metaBundle = Seomatic::$plugin->metaBundles->getGlobalMetaBundle($siteId);
         Seomatic::$previewingMetaContainers = false;
+        $hasErrors = false;
         if ($metaBundle) {
             /** @var array $scriptSettings */
             foreach ($scriptSettings as $scriptHandle => $scriptData) {
                 foreach ($metaBundle->metaContainers as $metaContainer) {
                     if ($metaContainer::CONTAINER_TYPE === MetaScriptContainer::CONTAINER_TYPE) {
                         $data = $metaContainer->getData($scriptHandle);
+                        /** @var MetaScript $data */
                         if ($data) {
                             /** @var array $scriptData */
                             foreach ($scriptData as $key => $value) {
@@ -922,10 +952,23 @@ class SettingsController extends Controller
                                     $data->$key = $value;
                                 }
                             }
+                            if (!$data->validate()) {
+                                $hasErrors = true;
+                            }
                         }
                     }
                 }
             }
+            if ($hasErrors) {
+                Craft::$app->getSession()->setError(Craft::t('app', "Couldn't save tracking settings due to a Twig error."));
+                // Send the redirect back to the template
+                Craft::$app->getUrlManager()->setRouteParams([
+                    'editedMetaBundle' => $metaBundle,
+                ]);
+
+                return null;
+            }
+
             Seomatic::$plugin->metaBundles->updateMetaBundle($metaBundle, $siteId);
 
             Seomatic::$plugin->clearAllCaches();