+++ src/services/MetaBundles.php
@@ -68,6 +68,30 @@ class MetaBundles extends Component
         'ogImageDescription',
     ];
 
+    const COMPOSITE_INHERITANCE_CHILDREN = [
+        'seoImage' => [
+            'metaBundleSettings.seoImageTransformMode',
+            'metaBundleSettings.seoImageTransform',
+            'metaBundleSettings.seoImageSource',
+            'metaBundleSettings.seoImageField',
+            'metaBundleSettings.seoImageIds',
+        ],
+        'ogImage' => [
+            'metaBundleSettings.ogImageTransformMode',
+            'metaBundleSettings.ogImageTransform',
+            'metaBundleSettings.ogImageSource',
+            'metaBundleSettings.ogImageField',
+            'metaBundleSettings.ogImageIds',
+        ],
+        'twitterImage' => [
+            'metaBundleSettings.twitterImageTransformMode',
+            'metaBundleSettings.twitterImageTransform',
+            'metaBundleSettings.twitterImageSource',
+            'metaBundleSettings.twitterImageField',
+            'metaBundleSettings.twitterImageIds',
+        ],
+    ];
+
     // Protected Properties
     // =========================================================================
 
@@ -606,9 +630,24 @@ class MetaBundles extends Component
             );
             // metaGlobalVars
             $attributes = $metaBundle->metaGlobalVars->getAttributes();
+
+            // Get a list of explicitly inherited values
+            $inherited = array_keys(ArrayHelper::remove($attributes, 'inherited', []));
             $emptyValues = array_fill_keys(array_keys(array_diff_key($attributes, $seoSettingsEnabledFields)), '');
+
+            // Nullify the inherited values
+            $emptyValues = array_merge($emptyValues, array_fill_keys($inherited, ''));
+            foreach ($inherited as $inheritedAttribute) {
+                foreach (self::COMPOSITE_INHERITANCE_CHILDREN[$inheritedAttribute] ?? [] as $child) {
+                    list ($model, $attribute) = explode('.', $child);
+                    $metaBundle->{$model}->$attribute = '';
+                }
+            }
+
             $attributes = array_merge($attributes, $emptyValues);
             $metaBundle->metaGlobalVars->setAttributes($attributes, false);
+
+
             // Handle the mainEntityOfPage
             if (!\in_array('mainEntityOfPage', $seoSettingsField->generalEnabledFields, false)) {
                 $metaBundle->metaGlobalVars->mainEntityOfPage = '';
@@ -620,7 +659,14 @@ class MetaBundles extends Component
             $metaBundle->metaSiteVars->setAttributes($attributes, false);
             // metaSitemapVars
             $attributes = $metaBundle->metaSitemapVars->getAttributes();
+
+            // Get a list of explicitly inherited values
+            $inherited = array_keys(ArrayHelper::remove($attributes, 'inherited', []));
             $emptyValues = array_fill_keys(array_keys(array_diff_key($attributes, $seoSettingsEnabledFields)), '');
+
+            // Nullify the inherited values
+            $emptyValues = array_merge($emptyValues, array_fill_keys($inherited, ''));
+
             $attributes = array_merge($attributes, $emptyValues);
             $metaBundle->metaSitemapVars->setAttributes($attributes, false);
         }